requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.4
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-filter==2.0.0
django-guardian==1.4.9
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.1.0
Pillow==5.4.1
PyJWT==1.7.1
pytz==2018.9
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
mysqlclient==1.3.13